🌱 split auth: extract provider into separate module
